refactor: split onboard starting point
